Change item time representation to int64.
The HN API page mentions only that it's a unix timestamp and links to
the corresponding wikipedia entry, which itself mentions that they are
typically signed, so that's what we're going to do.

+++ src/api/types.go
@@ -26,7 +26,7 @@ type Item struct {
 	By *string `json:"by"`
 
 	// Creation date of the item in Unix time.
-	Time *uint32 `json:"time"`
+	Time *int64 `json:"time"`
 
 	// The comment, story, or poll text in HTML.
 	Text *string `json:"text"`
